Work towards having pdp be a self-contained app.
The aim is to be able to run on a local workstation with no setup that
isn't declared in the project. No apache, no certs, no prior
assembling of js and css libs. mock data where necessary. Overview of
the change is that site-wide config moves into pdp-site/, and pdp/ is
our app. This is an adaptation of ACA practices, which don't use store
site-wide config except in directory travis-ci/ to aid in
testing. Still to be worked out is environment-specific settings, and
statics. restclients is configured to return mock data if user is
javerage, otherwise 404.

To run on a workstation one would do

 > REMOTE_USER=javerage python manage.py runserver